Lots of more code porting, mostly tests
