+++ janitor/src/main/resources/reference.conf
@@ -6,11 +6,11 @@ database {
   password = ${LEONARDO_DB_PASSWORD}
 }
 
-leonardo-pubsub {
-  topic-name = "leonardo-pubsub"
-}
-
 report-destination-bucket = "replace-me"
+
+leonardo-pubsub {
+  google-project = "test-project"
+}
 
 runtime-checker-config {
   path-to-credential = ${path-to-credential}
