+++ janitor/src/test/scala/com/broadinstitute/dsp/janitor/KubernetesClusterRemoverSpec.scala
@@ -1,18 +1,19 @@
 package com.broadinstitute.dsp
-package resourceValidator
+package janitor
 
 import cats.effect.IO
 import cats.mtl.Ask
 import com.broadinstitute.dsp.Generators._
+import com.broadinstitute.dsp.{CheckRunnerDeps, CronJobsTestSuite, KubernetesClusterToRemove, LeoPublisherDeps}
 import fs2.Stream
 import io.circe.Encoder
 import org.broadinstitute.dsde.workbench.google2.GooglePublisher
 import org.broadinstitute.dsde.workbench.google2.mock.{FakeGooglePublisher, FakeGoogleStorageInterpreter}
 import org.broadinstitute.dsde.workbench.model.TraceId
-import org.scalatest.flatspec.AnyFlatSpec
 import org.broadinstitute.dsde.workbench.openTelemetry.FakeOpenTelemetryMetricsInterpreter
+import org.scalatest.flatspec.AnyFlatSpec
 
-class KubernetesClusterRemoverSpec extends AnyFlatSpec with CronJobsTestSuite {
+final class KubernetesClusterRemoverSpec extends AnyFlatSpec with CronJobsTestSuite {
   it should "send DeleteKubernetesClusterMessage when clusters are detected to be auto-deleted" in {
     forAll { (clusterToRemove: KubernetesClusterToRemove, dryRun: Boolean) =>
       val dbReader = new FakeDbReader {
